requirements.txt
astronomer-providers==1.6.0